fix(supabase): points_to_yen の引数型を numeric に修正（sum(delta) と不一致でビュー作成失敗）
- sum(delta) は numeric を返すが points_to_yen(bigint) 定義で型解決できず
  admin_overview ビュー作成時に 42883 エラー → 引数を numeric に変更
- ローカルPostgresで setup_all.sql を全1211行エラーなく実行できることを確認
  （39テーブル/53関数/4ビュー、seed投入、admin_overview の円換算も動作）
- setup_all.sql を再生成

+++ supabase/migrations/0008_rate.sql
@@ -20,9 +20,9 @@ create or replace function public.point_yen_rate() returns int
   select (value #>> '{}')::int from public.app_config where key = 'point_yen_rate';
 $$;
 
--- ポイント→円換算ヘルパー
-create or replace function public.points_to_yen(p bigint) returns numeric
-  language sql stable as $$ select round(p::numeric / public.point_yen_rate()); $$;
+-- ポイント→円換算ヘルパー（sum(delta) は numeric を返すため引数も numeric で受ける）
+create or replace function public.points_to_yen(p numeric) returns numeric
+  language sql stable as $$ select round(p / public.point_yen_rate()); $$;
 
 -- ---------- XP をポイントから分離（活動量ベース） ----------
 alter table public.missions add column if not exists xp_reward integer not null default 50;
